\name{ralite-package}
\alias{ralite-package}
\docType{package}
\title{
A simple robust iterative web server for R
}
\description{
Provides a web server and development environment for R.
}
\details{
\tabular{ll}{
Package: \tab ralite\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-15\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey Horner <jeffrey.horner@gmail.com>
}
